Core: Navigation: remove SetWayPoints lambda. Tidy up variable names. Detect 'UniformPath' which means between every waypoint theres a 'close to average' distance. Incase the path is Uniform less likely to poke pathfinder when following the route.

+++ Core/GoalsComponent/PlayerDirection.cs
@@ -39,7 +39,7 @@ namespace Core
             SetDirection(desiredDirection, point, DefaultIgnoreDistance, _cts);
         }
 
-        public void SetDirection(float desiredDirection, Vector3 point, int ignoreDistance, CancellationTokenSource cts)
+        public void SetDirection(float desiredDirection, Vector3 point, float ignoreDistance, CancellationTokenSource cts)
         {
             float distance = playerReader.PlayerLocation.DistanceXYTo(point);
             if (distance < ignoreDistance)
